audio: Add OMAP_ENHANCEMENT to card name query
Primary audio HAL uses mixer_get_card_name() function (from
tinyalsa which is under OMAP_ENHANCEMENT) to dynamically find
the id of the McASP-based sound card.

The dynamic search of the card must also be dependent on
OMAP_ENHANCEMENT and provide a default card id if the flag is
not defined.

Change-Id: Id2c610e85d65bc5d559a0c9d1dda0359893a12fa
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>