audio: Multizone: Add support for voice call
Voice call is supported using two data pipes:

Uplink:

    PcmWriter <-- MonoPipe <-- PcmReader
       |                           |
     McASP7                      McASP3
       |                           |
    Bluetooth                     Mic

Downlink:

    PcmReader --> MonoPipe --> PcmWriter
       |                           |
     McASP7                      McASP3
       |                           |
    Bluetooth                   Speaker

Voice call reuses the PcmWriter associated with Cabin/Speaker that
is otherwise used for media. So, when the voice call starts any
ongoing media stream active on that writer has to be re-routed to
a null writer and routed back to the actual writer when the voice
call ends.

The null writer is a PcmWriter that writes data to a null output port.
The null port consumes the data at the same rate that the actual port
does. IOW, media streams going to the Cabin output (primary output)
during voice call will not be rendered on the actual hardware output.

Change-Id: I7856b67ab1cf788ea2f2cb251d9097e57e487cb5
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>