audio: Multizone: Bluetooth slave mode workaround
The PCM ports used for Bluetooth SCO are in slave mode in Jacinto6
EVM board. So the BT side masters the BCLK and FSYNC. The BT side
could cut the BCLK and FSYNC while the ALSA PCM ports are not closed,
which leads to blocking read() and write() calls. They would eventually
timeout and cause an I/O error.

The workaround consists of switching the BT DAI link to slave mode
(McASP is master, BT is slave) while leaving the call, it guarantees
that the BCLK and FSYNC are present as they are generated from an
internal source, the actual rate of the clock is irrelevant.
For the new source to take effect, the ALSA PCM ports used for BT are
explicitly stopped.

Change-Id: Ifadb8b1348529d404d06d4b043c5fe686f85a49a
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>