audio: Multizone: Dynamic JAMR3 support
JAMR3 board may not always be present in the system, the port
usage is being changed dynamically to:

- JAMR3 is present
   o Cabin   : port 1, slots 0 & 1
   o Mic     : port 1, slot 2
   o Back Mic: port 1, slot 3

- JAMR3 is not present
   o Cabin   : port 0, slots 0 & 1
   o Mic     : port 0, slots 0 & 1

Above port allocation allows:
- Relinquish McASP3 (port 0) when JAMR3 is connected. This allows
  DSP Radio to own that McASP instance for its rendering.
- Hardware setups that don't have JAMR3 fall back to the port 0
  that is present in the main Jacinto6 EVM board (CPU board).

The default hardware setup is assumed to have JAMR3 connected.
When it's not connected, the property "persist.audio.use_jamr"
must be set to "0" or "false" and system has to be rebooted
(or media server restarted). This is meant to be a development
option.

Change-Id: Icc5399b370e2247a1a0d5a967dcd47e0ef4947ac
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>