audio: Multizone: Don't Round to 16 Input Buffer Size
The input buffer size returned from AudioStreamIn::getBufferSize()
and AudioHwDevice::getInputBufferSize() was being rounded to a
multiple of 16 with a note that it is required by AudioFlinger,
but AudioFlinger has no such requirement for input buffer size
(only output buffer size). Rounding to 16 caused an issue for the
multizone audio when certain sample rates were used. So, the
rounding is removed as it is not needed.

Change-Id: I8d5010f5bab7eb780868b47899ea0115fa8c1c4b
Signed-off-by: Angela Stegmaier <a0866189@ti.com>