audio: Legacy: BT SCO voice call support
The Bluetooth SCO voice call has the following data flow:

Uplink:

       +----------- SRC -----------+
       |                           |
     McASP7                      McASP3
       |                           |
    Bluetooth                     Mic

Downlink:

       +----------- SRC -----------+
       |                           |
     McASP7                      McASP3
       |                           |
    Bluetooth                   Speaker

The voice call uses the ALSA PCM devices that are normally used
by the primary output and input. The primary output is switched
to a null/dummy sink when the voice call starts, and switched
back to the ALSA devices when the call ends.

The null/dummy sink consumes the data at the same rate than an
actual PCM device does, but does not render anything to an output
device (e.g. speaker).

Change-Id: I63a41de96c4ed34b60aab9240d10ad83c06c69ac
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>