audio: Legacy: Add DRA7xxEVM to supported cards
There is an audio driver for the DRA7 EVM that supports a
single PCM device (McASP3).  The name of the card created
by this driver is "DRA7xxEVM", so this name has to be added
to the cards supported by this Audio HAL.

The IN_CALL mode is supported in the Audio HAL but won't
work with the "DRA7xxEVM" card because the PCM device used
for BT is not implemented by the driver.

Change-Id: Iacee36175cd4c90e74fce57f41cc6ca7c07291b9
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>